requirements.txt
python-novaclient!=7.0.0,>=6.0.0 # Apache-2.0
python-neutronclient>=5.1.0 # Apache-2.0
python-keystoneclient>=3.8.0 # Apache-2.0
python-glanceclient>=2.5.0 # Apache-2.0
python-heatclient>=1.6.1 # Apache-2.0
python-cinderclient
ansible>=2.1.0,<2.4
wrapt>=1.7.0 # BSD License
scp
cryptography!=1.3.0,>=1.0 # BSD/Apache-2.0
